+++ packages/brick_sqlite_generators/lib/src/sqlite_serialize.dart
@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
-import 'package:analyzer/dart/element/element.dart' show ClassElement;
-import 'package:analyzer/dart/element/type.dart' show DartType;
-import 'package:brick_sqlite_abstract/db.dart' show InsertTable;
+import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/type.dart';
+import 'package:brick_sqlite_abstract/db.dart' show InsertTable, InsertForeignKey;
 import 'package:brick_sqlite_abstract/sqlite_model.dart';
 import 'package:source_gen/source_gen.dart' show InvalidGenerationSourceError;
 import 'package:brick_sqlite_generators/src/sqlite_serdes_generator.dart';
@@ -54,11 +54,16 @@ class SqliteSerialize<_Model extends SqliteModel> extends SqliteSerdesGenerator<
     }
 
     final primaryKeyByUniqueColumns = generateUniqueSqliteFunction(uniqueFields);
+    final afterSaveCallbacks = unignoredFields.where((f) {
+      final checker = checkerForType(f.type);
+      return checker.isIterable && checker.isArgTypeASibling;
+    }).map(_saveIterableAssociationFieldToJoins);
 
     return [
       'final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {${fieldsToColumns.join(',\n')}};',
       primaryKeyByUniqueColumns,
-      "final String tableName = '$tableName';"
+      "final String tableName = '$tableName';",
+      "Future<void> afterSave(instance, {provider, repository}) async {$afterSaveCallbacks.join('\n')}"
     ];
   }
 
@@ -89,37 +94,6 @@ class SqliteSerialize<_Model extends SqliteModel> extends SqliteSerdesGenerator<
       // Iterable
     } else if (checker.isIterable) {
       final argTypeChecker = checkerForType(checker.argType);
-
-      if (checker.isArgTypeASibling) {
-        // Iterable<Future<SqliteModel>>
-        if (checker.isArgTypeAFuture) {
-          return '''jsonEncode(
-            (await Future.wait<int>($fieldValue
-              ?.map(
-                (s) async => (await s)?.${InsertTable.PRIMARY_KEY_FIELD} ?? await provider?.upsert<${checker.unFuturedArgType}>((await s), repository: repository)
-              )
-              ?.toList()
-              ?.cast<Future<int>>()
-              ?? []
-            )).where((s) => s != null).toList().cast<int>()
-          )''';
-
-          // Iterable<SqliteModel>
-        } else {
-          final instanceAndField = wrappedInFuture ? '(await $fieldValue)' : fieldValue;
-
-          return '''jsonEncode(
-            (await Future.wait<int>($instanceAndField
-              ?.map((s) async {
-                return s?.${InsertTable.PRIMARY_KEY_FIELD} ?? await provider?.upsert<${checker.unFuturedArgType}>(s, repository: repository);
-              })
-              ?.toList()
-              ?.cast<Future<int>>()
-              ?? []
-            )).where((s) => s != null).toList().cast<int>()
-          )''';
-        }
-      }
 
       // Iterable<enum>
       if (argTypeChecker.isEnum) {
@@ -201,6 +175,41 @@ class SqliteSerialize<_Model extends SqliteModel> extends SqliteSerdesGenerator<
 
       return results.first['${InsertTable.PRIMARY_KEY_COLUMN}'];
     }""";
+  }
+
+  String _saveIterableAssociationFieldToJoins(FieldElement field) {
+    final annotation = fields.annotationForField(field);
+    var checker = checkerForType(field.type);
+    final fieldValue = serdesValueForField(field, annotation.name, checker: checker);
+
+    final wrappedInFuture = checker.isFuture;
+    if (wrappedInFuture) {
+      checker = checkerForType(checker.argType);
+    }
+
+    // Iterable<Future<SqliteModel>>
+    final insertStatement =
+        'INSERT OR REPLACE INTO `${InsertForeignKey.joinsTableName(field.name, localTableName: fields.element.name)}` (`${InsertForeignKey.foreignKeyColumnName(fields.element.name)}`, `${InsertForeignKey.foreignKeyColumnName(checker.unFuturedArgType.getDisplayString())}`)';
+    var siblingAssociations = fieldValue;
+    var upsertMethod =
+        '(await s)?.${InsertTable.PRIMARY_KEY_FIELD} ?? await provider?.upsert<${checker.unFuturedArgType}>((await s), repository: repository)';
+
+    // Iterable<SqliteModel>
+    if (!checker.isArgTypeAFuture) {
+      siblingAssociations = wrappedInFuture ? '(await $fieldValue)' : fieldValue;
+      upsertMethod =
+          's?.${InsertTable.PRIMARY_KEY_FIELD} ?? await provider?.upsert<${checker.unFuturedArgType}>(s, repository: repository)';
+    }
+
+    return '''
+          await Future.wait<int>($siblingAssociations
+            ?.map((s) async => ($upsertMethod)?.then((id) => 
+              ${InsertTable.PRIMARY_KEY_FIELD} != null 
+                ? provider?.rawExecute('${insertStatement}', [${InsertTable.PRIMARY_KEY_FIELD}, id]) 
+                : null
+            ))
+          );
+        ''';
   }
 
   String _finalTypeForField(DartType type) {